+++ test/UserRepository/HtpasswdFactoryTest.php
@@ -25,16 +25,17 @@ class HtpasswdFactoryTest extends TestCase
 
     public function testInvokeWithMissingConfig()
     {
-        $this->container->has('config')->willReturn(false);
-        $this->container->get('config')->shouldNotBeCalled();
+        // We cannot throw a ContainerExceptionInterface directly; this
+        // approach simply mimics `get()` throwing _any_ exception, which is
+        // what will happen if `config` is not defined.
+        $this->container->get('config')->willThrow(new InvalidConfigException());
 
         $this->expectException(InvalidConfigException::class);
-        $htpasswd = ($this->factory)($this->container->reveal());
+        ($this->factory)($this->container->reveal());
     }
 
     public function testInvokeWithEmptyConfig()
     {
-        $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn([]);
 
         $this->expectException(InvalidConfigException::class);